refactor(api): move article aggregate to domain
